The artifact and group name have been renamed
Package has been renamed

Window has a title.
