+++ pom.xml
@@ -4,13 +4,15 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>org.example</groupId>
-    <artifactId>diagnosticsanalyzer</artifactId>
+    <groupId>com.hazelcast</groupId>
+    <artifactId>diagnostics-analyzer</artifactId>
     <version>1.0-SNAPSHOT</version>
+
     <properties>
         <maven.compiler.target>11</maven.compiler.target>
         <maven.compiler.source>11</maven.compiler.source>
     </properties>
+
     <dependencies>
         <dependency>
             <groupId>com.jidesoft</groupId>
@@ -23,7 +25,6 @@
             <artifactId>jfreechart</artifactId>
             <version>1.5.3</version>
         </dependency>
-
     </dependencies>
 
 </project>
