+++ src/main/java/com/hazelcast/diagnostics/HeartbeatPane.java
@@ -1,4 +1,4 @@
-package com.hazelcast.tricorder;
+package com.hazelcast.diagnostics;
 
 import javax.swing.*;
 import java.util.ArrayList;
