+++ src/main/java/com/hazelcast/diagnostics/OperationsPlane.java
@@ -1,4 +1,4 @@
-package com.hazelcast.tricorder;
+package com.hazelcast.diagnostics;
 
 import org.jfree.chart.ChartFactory;
 import org.jfree.chart.ChartPanel;
@@ -12,27 +12,26 @@ import org.jfree.data.time.TimeSeriesCollection;
 
 import javax.swing.*;
 import java.text.SimpleDateFormat;
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Iterator;
 import java.util.Map;
 
-public class MemoryPane {
+public class OperationsPlane {
 
     private final JPanel component;
     private final JFreeChart chart;
     private final ChartPanel chartPanel;
     private final TimeSeriesCollection collection;
-    private Collection<InstanceDiagnostics> diagnosticsList = new ArrayList<>();
+    private Collection<InstanceDiagnostics> diagnosticsList;
     private long startMs = Long.MIN_VALUE;
     private long endMs = Long.MAX_VALUE;
 
-    public MemoryPane() {
+    public OperationsPlane() {
         collection = new TimeSeriesCollection();
         chart = ChartFactory.createTimeSeriesChart(
-                "Memory Usage",
+                "Operation Throughput",
                 "Time",
-                "Memory Usage",
+                "Throughput",
                 collection,
                 true,
                 true,
@@ -55,21 +54,32 @@ public class MemoryPane {
 
     public void update() {
         collection.removeAllSeries();
+        if (diagnosticsList == null) {
+            return;
+        }
 
         for (InstanceDiagnostics diagnostics : diagnosticsList) {
-            Iterator<Map.Entry<Long, Number>> iterator = diagnostics.metricsBetween("[metric=runtime.usedMemory]", startMs, endMs);
+            Iterator<Map.Entry<Long, Number>> iterator = diagnostics.metricsBetween("[unit=count,metric=operation.completedCount]", startMs, endMs);
 
             if (!iterator.hasNext()) {
                 continue;
             }
 
             TimeSeries series = new TimeSeries(diagnostics.getDirectory().getName());
+            long previousMs = 0;
+            long previousCount = 0;
             while (iterator.hasNext()) {
                 try {
                     Map.Entry<Long, Number> entry = iterator.next();
-                    long ms = entry.getKey();
-                    Long value = entry.getValue().longValue();
-                    series.add(new FixedMillisecond(ms), value);
+                    long currentMs = entry.getKey();
+
+                    long count = entry.getValue().longValue();
+                    long delta = count-previousCount;
+                    long durationMs = currentMs - previousMs;
+                    double throughput = (delta * 1000d) / durationMs;
+                    series.add(new FixedMillisecond(currentMs), throughput);
+                    previousMs = currentMs;
+                    previousCount=count;
                 } catch (SeriesException e) {
                     System.err.println("Error adding to series");
                 }
