+++ src/main/java/com/hazelcast/diagnostics/SlowOperationsPane.java
@@ -1,4 +1,4 @@
-package com.hazelcast.tricorder;
+package com.hazelcast.diagnostics;
 
 import javax.swing.*;
 import java.util.ArrayList;
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.TreeMap;
 
-public class ConnectionPane {
+public class SlowOperationsPane {
 
     private final JComponent component;
     private final JTextArea textArea;
@@ -16,7 +16,7 @@ public class ConnectionPane {
     private long startMs = Long.MIN_VALUE;
     private long endMs = Long.MAX_VALUE;
 
-    public ConnectionPane() {
+    public SlowOperationsPane() {
         this.textArea = new JTextArea();
         textArea.setEditable(false);
         this.component = new JScrollPane(textArea);
@@ -39,7 +39,7 @@ public class ConnectionPane {
 
         TreeMap<Long, List<String>> treeMap = new TreeMap();
         for (InstanceDiagnostics diagnostics : diagnosticsList) {
-            Iterator<Map.Entry<Long, String>> iterator = diagnostics.between(InstanceDiagnostics.TYPE_CONNECTION, startMs, endMs);
+            Iterator<Map.Entry<Long, String>> iterator = diagnostics.between(InstanceDiagnostics.TYPE_SLOW_OPERATIONS, startMs, endMs);
 
             if (!iterator.hasNext()) {
                 continue;
