+++ src/main/java/com/hazelcast/diagnostics/SystemPropertiesPane.java
@@ -1,4 +1,4 @@
-package com.hazelcast.tricorder;
+package com.hazelcast.diagnostics;
 
 import javax.swing.*;
 import javax.swing.table.DefaultTableModel;
